+++ src/applications/static/single_threaded/pagerank.h
@@ -1,7 +1,7 @@
 #ifndef KAPPA_PAGERANK_H
 #define KAPPA_PAGERANK_H
 
-const double DEFAULT_DAMPING_FACTOR = 0.85;
+const float DEFAULT_DAMPING_FACTOR = 0.85;
 const int DEFAULT_MAX_ITERATIONS = 20;
 
 void init_pr_values(Digraph*);
