+++ src/experiments/utils/dataset_split.h
@@ -1,7 +1,7 @@
 #ifndef KAPPA_DATASET_SPLIT_H
 #define KAPPA_DATASET_SPLIT_H
 
-unsigned long number_of_lines(std::string);
-std::vector<unsigned long> dataset_split_by_size(unsigned long, unsigned long, unsigned long);
+unsigned int number_of_lines(std::string);
+std::vector<unsigned int> dataset_split_by_size(unsigned int, unsigned int, unsigned int);
 
 #endif //KAPPA_DATASET_SPLIT_H
